Move most engine code to a separate file
